+++ CHANGELOG.md
@@ -2,12 +2,12 @@
 
 Mapbox welcomes participation and contributions from everyone.
 
-# main
+# 10.11.2
 ## Bug fixes 🐞
-* Fix missing terrain on some GPUs (e.g. Mali). ([2036](https://github.com/mapbox/mapbox-maps-android/pull/2036)) 
+* Fix missing terrain on some GPUs (e.g. Mali). ([2036](https://github.com/mapbox/mapbox-maps-android/pull/2038)) 
 
 ## Dependencies
-* Update gl-native to v10.11.2. ([2036](https://github.com/mapbox/mapbox-maps-android/pull/2036))
+* Update gl-native to v10.11.2. ([2036](https://github.com/mapbox/mapbox-maps-android/pull/2038))
 
 # 10.11.1 February 21, 2023
 ## Bug fixes 🐞